webusb work great with linux
