+++ src/device/usbd.c
@@ -136,8 +136,8 @@ static usbd_class_driver_t const usbd_class_drivers[] =
       .init             = vendord_init,
       .reset            = vendord_reset,
       .open             = vendord_open,
-      .control_request  = NULL, // tud_vendor_control_request_cb
-      .control_complete = NULL, // tud_vendor_control_complete_cb
+      .control_request  = tud_vendor_control_request_cb,
+      .control_complete = tud_vendor_control_complete_cb,
       .xfer_cb          = vendord_xfer_cb,
       .sof              = NULL
   },
